requirements.txt
pillow==9.3.0
pytemperature==1.1
requests==2.28.1
